% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_estimation.R
\name{g.estimation}
\alias{g.estimation}
\title{G-estimation method}
\usage{
g.estimation(data, y, a, q.model, model.family, estimate = "diff")
}
\arguments{
\item{estimate}{Estimate difference or ratio. Defaults to "diff"}
}
\description{
Function calculates a "causal" estimate of a binary
treatment/exposure (a) on an outcome (y) using simple substitution
(g-estimation). It can calculate the risk difference or risk ratio.
}
\examples{
g.estimation()
}
\keyword{gestimation}
