% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_formula.R
\name{g.formula}
\alias{g.formula}
\title{G-formula method}
\usage{
g.formula(data, y, a, q.model, model.family, estimate = "diff")
}
\arguments{
\item{estimate}{Estimate difference or ratio. Defaults to "diff"}
}
\description{
Function calculates a "causal" estimate of a binary
treatment/exposure (a) on an outcome (y) using simple substitution
(g-formula). It can calculate the risk difference or risk ratio.
}
\examples{
g.formula()
}
\keyword{g-formula}
