% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw_estimation.R
\name{iptw.estimation}
\alias{iptw.estimation}
\title{IPTW-estimation method}
\usage{
iptw.estimation(data, y, a, w, estimate = "diff")
}
\arguments{
\item{estimate}{Estimate difference or ratio. Defaults to "diff"}
}
\description{
Functions calculates a "causal" estimate of a binary
treatment/exposure (a) on an outcome (y) using inverse probability of treatment
weighting (IPTW). It can calculate the risk difference or risk ratio.
}
\examples{
iptw.estimation()
}
\keyword{iptw}
