% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw.R
\name{iptw}
\alias{iptw}
\title{IPTW method}
\usage{
iptw(data, y, a, w, estimate = "diff")
}
\arguments{
\item{estimate}{Estimate difference or ratio. Defaults to "diff"}
}
\description{
Functions calculates a "causal" estimate of a binary
treatment/exposure (a) on an outcome (y) using inverse probability of treatment
weighting (IPTW). It can calculate the risk difference or risk ratio.
}
\examples{
iptw()
}
\keyword{IPTW}
\keyword{inverse-probability-of-treatment}
\keyword{iptw,}
\keyword{weight,}
